sample factory for all samplers
